LuaJIT: Use LuaJIT PRNG instead of Redis one.
A random number implementation was introduced to Redis to make sure
it outputs the same sequence (for the same seed) in every arch.  Redis
overrides math.randdom with this implementation.

LuaJIT has a suitable PRNG so we use that instead.

The Redis PRNG was only used for scripting, so this changeset removes it.